refactor(review): timePL.derivers から pure derivation と types を domain へ分離
UI formatter / Tailwind class helper と混在していた timePL.derivers を整理。
pure derivation と関連 types を features/review/domain/timePL/ に切り出し、
UI 表示用 helper は components/time-pl/data/timePL.presentation.ts に分離。

- domain/timePL/derivers.ts: deriveAccuracy / deriveStatement / deriveWaterfall /
  deriveBarComparison / deriveBreakEven / deriveBalanceSheet / getAccuracyStatus
- domain/timePL/types.ts: TimePLInput / StatementViewData 等の全 type
- data/timePL.presentation.ts: formatMinutesDuration / formatVariance /
  getVarianceColor / getAccuracyColors
- テストも pure 部分 (domain) と presentation 部分に分割
- 仕様 / UI / DB 変更なし、derivation ロジックは無改変
